Auto-restart worker processes when they die
On prod, we have a primary process and two worker processes. If a worker
process dies (such as by running out of memory), we previously wouldn't
notice this, so if a server lost a worker process this way it would
silently lose half its capacity. (Then if it lost its other worker, it
would stop being able to service requests at all, and the load balancer
would stop sending it any requests.)

Co-Authored-By: Will Howard <w.howard256+github@gmail.com>