The tab navigation footer triggers on mouse down, not mouse up
Fixes an issue in the `HashLink` component where it failed to
`preventDefault` the `onClick` event. This doesn't matter for links in
parts of the page that will no longer exist after navigation (so wasn't
an issue before), but is an issue for links in the navigation footer,
which persists across pages.