Concepts and Community pages have a navigation footer
Makes it so the LW Concepts and Community pages (which are linked to from
the bottom-bar) have bottom-bars themselves.

Disentangles the concept of "having standalone navigation", which
referred to both having left-side ToC-style navigation options at larger
breakpoints, and having a bottom-bar at smaller breakpoints. That didn't
work for the Concepts and Community pages, which have their own tables
of contents and want the bottom-bar without the ToC-style page
navigation.