Better mobile footnotes
On large screens, footnotes render as sidenotes. On smaller screens,
footnotes have a preview when you hover with a mouse. On _touch_
screens, however, you can't hover, so the only interaction left was that
clicking on a footnote would scroll to it in the footer, which kind
works but isn't great.

Change it so that when you tap on a footnote on mobile ("mobile" defined
according to the `bowser` library, should pretty closely match devices
with touchscreens), you get the footnote contents in a modal dialog with
a clickaway, rather than scrolling.