Remove an invalid JSS rule that caused a console-log warning on startup
@property is a CSS feature for controlling the behavior (inheritance,
default value, typecheck) of custom CSS properties. It was used in
TopPostsPage, but JSS isn't aware of it, so it produced a warning on
startup. In an unsuccessful attempt to fix this in
c3214b025117ae7faddaf5ee57bf3b89b184192d, the same @property
specification was added to the `rawCSS` section of createThemeDefaults,
but with CSS syntax errors in it.

Remove the version that creates a warning, and fix the syntax of the
version that doesn't.