Make Google Maps API loading use `async`
The Google Maps API is used for autocompleting form location inputs,
which can appear on the front page in the widget that suggests you sign
up for nearby-meetup notifications. This produced a console-log warning
saying that it was loading in non-async mode, which can impact
performance. Change the loading method to be async (it's already
correctly waiting for a callback, so this is safe).