Remove some things from `Utils` that didn't need to be there
Monkeypatching onto the `Utils` object is a mechanism for code in `lib`
to call functions that are in the `server` directory (so they can't be
imported). Some of the functions no longer need to be there, since all
the code that called them that way is gone or has been moved into
`server`.