Fix the Add Posts button on tag pages not updating client-side
The Add Posts button on tag pages gives you a search box for posts, and
adds the tag to them. There was some code there to do a client-side
update, so that when you tag a post this way it shows up in the list of
posts without a refresh, but this didn't work due to a fragment issue.