Support cross-environment imports with stub files
Previously, importing code in `packages/lesswrong/client` or
`packages/lesswrong/server` from code in `components` or `lib` wouldn't
work, because that would lead to server-specific code in the client
bundle, or vise versa. This lead to some fairly awkward hacks involving
`require()` at runtime, or server code assigning things into global
variables so that shared code can retrieve it.

This PR makes it so that shared code can import from the client- and
server-directories, so long as (a) it uses "@/client" or "@/server" (not
"../../client" or "../../server"), and (b) there is a file with the same
name in packages/lesswrong/stubs. When the client bundle tries to import
server code, it will get the version from the stubs directory; and
similarly when the server tries to import client code.

Note that while stub files are typechecked, we don't typecheck twice;
the build will fail if a stub file is required but not present, or
doesn't export stub versions of the function that's imported, but if a
stub file exports things of different types than what it's replacing,
that isn't necessarily detected.