Reduced size of user menu; added dialogue button to user profile
On LW removed buttons from the long user menu that were not getting any use. We still need some way to make dialogues, so every user profile has a 'dialogue' button to start a dialogue with them. Also fixed a bug in the menu where the bottom of the g of 'log out' was clipped. Forum gated, shouldn't affect EA Forum menu.

Co-Authored-By: Jim Babcock <jim@jimrandomh.org>