Remove http->https redirect from nginx config
When I added Cloudfront as a caching proxy in front of an LW development
staging server (baserates.org), it resulted in a redirect loop, because
some request forwarding within AWS infra was http, even though the
external connection was https. Looking at how our http->https redirects
work in general, this redirector seems to be redundant; on all of LW,
AF, and EA Forum, clients get redirected from http->https before the
request gets far enough for nginx to see it; we can tell that the https
redirection that's triggering is not this one, because that https
redirection uses a 307 status code rather than 301.