Add vite server option to build script
Adds a vite-server mode, in which the backend and SSR works as before,
but the frontend is provided by vite and has hot module reloading.
Addresses various issues that made our code incompatible with vite.

Time to interactivity from a fresh pageload with vite is drastically
worse than it is with traditional-esbuild. In exchange for this, you get
a lot of incrementality once a tab is ready.

The vite server serves the frontend, and passes along all SSR requests
and graphql queries to a separate backend server which we build as
before using esbuild. When you start a fresh pageload, it goes to a vite
server running in the build process, which forwards the request to a
backend server where it is SSR'ed, then modifies the result to load vite
scripts at the end.

Changes that affect the client are handled by vite, which will either
perform hot-module reloading (if the changed file is a component file
which also exports the return value of `registerComponent` as a default
export), or perform a tab refresh (which is reasonably fast because it
doesn't involve restarting the backend server).

Currently, the backend server is on port 3000 and the vite frontend is
on port 3010. A future commit will put the vite server on 3000, and
overlap backend servers using multiple ports to allow rebuilding the
backend seemlessly.

Styles are still provided by a static stylesheet. I tried switching over
to `withStyles` HoCs, and this turned out to be pretty broken; while
there is code for injecting JSS styles into the page header that we can
dig out of the git history, the underlying library doesn't have anything
analogous to the `stylePriority` option that we use during static-style
generation, and this causes incorrect style prioritization and visual
glitches all over the front page.