Overlap backend servers while using vite
When --vite is passed to the build script, start a new server and wait
for it to be ready before stopping the old one. This makes it so that if
you make changes that are handled by HMR on the frontend, there won't be
any time period in which there isn't a server available for SSRs and
graphql requests.