Add useStyles hook; HMR updates styles
Previously, we used a merged static stylesheet generated by the server
during startup, with JSS's builtin style-mounting as a fallback. Over
time, we came to rely on features that our static stylesheet had which
were missing from the JSS library, in particular the `stylePriority`
option. This fallback case was only used what the /allStyles failed to
load, ie approximately never, so that wasn't an issue.

With hot module reloading, however, we need to manage styles from each
of our components separately rather than in one big merged stylesheet.
With this commit, we make HMR able to support styles, take control of
our own style mounting, and introduce a `useStyles` idiom (as an
alternative to passing `styles` to `registerComponent` and having
`classes` passed in by an HoC).

As a hopefully-temporary limitation, in order to get HMR of styles, you
must use the new `useStyles` hook.