Don't fail when converting a page that has a LaTeX error in it
In mjPagePromise, our error handling would cause the conversion as a
whole to throw an exception. In the context of Arbital import, this
meant that creating revisions with LaTeX errors in them would fail, and
those revisions would be missing from the history.

Rather than reject the promise (throw an error) in this case, insert an
error message into the document in the spot where the invalid LaTeX
formula was, as described in the mathjax-node-page documentation.