Make CkEditor able to load under Vite
This commit makes CkEditor able to load under vite. To do this we add
@ckeditor/vite-plugin-ckeditor5, which is marked as deprecated in npm
because it's supposedly not necessary anymore (but we empirically
determined that it was necessary).

We also rename all SVGs in our custom plugins to have names starting
with ckeditor5-*, because we were getting crash-on-starts related to SVG
loading, and when we looked inside vite-plugin-ckeditor5 we found that
it was checking filenames and only handling SVG files if they matched
that pattern.

Note that this commit also exposes our CkEditor plugins to stricter
typescript settings than they've had previously, which introduces a lot
of type errors. There are also a large number of warnings from vite of
the form:
    ```
    Assets in public directory cannot be imported from JavaScript.
    If you intend to import that asset, put the file in the src directory, and use /src/lesswrong-editor/node_modules/@ckeditor/ckeditor5-ui/theme/icons/color-tile-check.svg instead of /public/lesswrong-editor/node_modules/@ckeditor/ckeditor5-ui/theme/icons/color-tile-check.svg.
    ````
which is a symptom of `lesswrong-editor` being in `public. This doesn't
appear to be actually breaking anything, but the warning-spam is a bit
much and will need to be addressed.

Co-Authored-By: Ollie Etherington <ollie@etherington.xyz>