Greatly improve the non-HMR vite update experience
This makes it so that most of the time, editing a component that doesn't
meet vite's HMR-eligibility rules will still give you a fast refresh.

Before, editing a non-HMR component would rerun the component file (in
many cases successfully making the change visible), then trigger a full
refresh which would be served as an SSR by the previous version
(reverting the change), then run the updated Javascript (re-applying the
change, but with an SSR mismatch warning).

In this version, we suppress the full-page refresh triggered by vite
(using a horrible hack), and replace it with a one-time full
invalidation of the React tree. If the thing that you changed was a
React component that passes through registerComponent, or was a JSS
style that gets passed to registerComponent or defineStyles, then this
will successfully display your change in under a second. However, any
React state on the page held in useState, forms, etc will be lost,
similar to what happens if you did a full page refresh.

If the thing that you changed *wasn't* a component or JSS, then you can
get strange effects. In particular, imports in other files may continue
pointing to old versions of things, data saved in global variables and
other places may point to old versions, there may be mixes of old and
new versions causing equality-tests that should pass to instead fail,
etc. All such issues will resolve if you do a manual hard refresh.