Restore voteability of Elicit predictions
Previuosly, we had an API integration with Elicit, where you would
create questions on their site and we would display them as post embeds
with a graph of the distribution of votes, and you could click to vote.
Then they shut down that API, so we did a one-time import of all the
embedded questions and their votes, so that we would still be able to
display them in posts that used this feature. But we didn't implement
the ability to vote, without the cross-site API, or to create questions.

"Voteable predictions" are a nice feature to have in general, and also
it's a feature Arbital had and we're aiming for Arbital feature-parity.
So, this commit is a start on bringing that back. This doesn't (yet) let
you create new questions to predict, but does restore the ability to
predict for questions that already exist.