Fix missing users on the community map, and drastically reduce download-size of /community
The user-map-markers query on /community had a limit:500, but there are
more than that in the DB, so some users would be missing (and adding
yourself to the map would appear to fail). Fix that. Also, the download
size of the page with the map was pretty chonky (4.3MB), because of the
large list of users; solve that by trimming the UsersMapEntry fragment.
In particular, the mapLocation field was a Google Maps API object with a
bunch of random metadata about the location, but we only need the lat
and lng, so make a resolver-only field that returns a minimal version of
the location.