Improve memory usage of build.ts
Apparently running scripts with ts-node loads the typescript compiler
and a whole bunch of types, adding ~1GB to the memory usage of build.ts.
If we add --transpile-only, it doesn't. We get our typechecking by
adding it to tsconfig-shared.json instead, which causes it to be
typechecked at the same time as the rest of the codebase.