Fix crash when pasting LaTeX from a post/comment into CkEditor
Changes the CkEditor upcast for <span class="math-tex">, which is the
case that you get when copy/pasting from content that has had mathjax
run on it and is displayed in a ContentItemBody. Before, this would
crash; now it copies the LaTeX successfully.

I didn't update the CkEditor bundle version for this, and believe
changing the bundle version is unnecessary, because the crash always
happens client-side, and the behavior when not-crashing is
indistinguishable, so the version of the bundle that runs in the cloud
is effectively not different. I have tested that a collab-editor session
with the fixed client bundle and an unchanged bundle version appears to
work when pasting LaTeX from a post (the case that previously crashed).