Fix CkEditor side-comment highlight covering up adjacent lines in dialogues
In CkEditor dialogues, the font is such that the space reserved for
ascenders/descenders is large relativ to the line height. This interacts
badly with CkEditor side-comments, which add a yellow background
highlight to text ranges; the background highlight of each line covers
part of the line above it.

The ideal fix would be to create a stacking context and use z-indexes to
put all of the text on top of all of the background, as in this
stackoverflow thread: https://stackoverflow.com/a/76044890/274472 .
However, in order to do this we would need an extra wrapper span inside
the .ck-comment-marker span, and that's coming from a part of CkEditor
that we don't control, so we have to make do with a less-clean solution.
So instead we shrink the highlight vertically by 3px on the top and 3px
on the bottom, which reduces the overlap to only cutting off the bottoms
of descenders rather than cutting off more of the character, and we make
the highlight a higher-intensity color with transparency rather than a
lower-intensity color at full opacity, so that the descenders can show
through when they're in the overlap region.