PostsListBase only includes tag metadata, not preview text
Modifies the PostsListBase fragment to only include the metadata for
tags on the post, not text for a hover preview, moving the hover-preview
text to PostsDetails. Update type annotations in other places
accordingly. Makes FooterTag have with-preview and non-preview variants,
and makes TruncatedTagsList do a separate query to get tag previews
(this covers the case where you preload the hoverover for a post, and
don't want to download descriptions of tags on that post until after the
hoverover is opened.)