Fix a subtle memory leak caused by putting a resolverContext into async storage
Poking around in memory dumps, I found that a bounded (but significant)
number of `resolverContext` instances were being leaked on startup,
which is a problem because `resolverContext` includes a loader-cache
which is meant to be ephemeral, and contains a signficant amount of
data. The issue is that the internals of the `dataloader` library hold
onto a resolved promise, which captures the `Context` that the
`DataLoader` class was constructed in, which captures all associated
async local storage.

The only thing the `resolverContext` in `AsyncLocalStorageContext` was
being used for was as a source of a `trace_id`, put the same `trace_id`
is also available via a different, smaller async local storage variable.
I checked that the values from these two sources matched, then removed
`resolverContext` from AsyncLocalStorageContext.