Throw an error if reposting with a duplicate title
We've recently had a lot of errors where posts are getting duplicated
in very quite succession. Some of these are user error, some may be a
technical issue. In either case, this should prevent the repost by not
allowing users to publish a post if it has the same title as another
post they created in the last hour.