Fix the color of the dismiss button on spotlight items, while in dark mode
This was using palette colors which invert in dark mode (normally
correct), but overlayed on top of an image (which doesn't invert). The
result wasn't illegible or anything, but looked kind of ugly.

I forum-gated this change because there was already some forum-gating on
the `color` prop for `closeButton`, and I wasn't testing the
interaction.