Fix spurious cursor:pointer styles on hidden post-item triple dot menu
When you're logged out, the triple-dot menu next to post items is hidden
(because the stuff inside the menu is all only relevant for logged-in
users). But the cursor:pointer style was still there on the spot where
the menu would be, even when the menu was omitted. Move styles around to
fix that.