Add a "skip attributions" option to revisions
This adds a moderator/admin-only context menu to revisions on the tag
history page, with an "Exclude this revision from attributions" option.
If set, then when we estimate the size of contributions for the "Written
by" line on the tag page, and when we decide which sections to highlight
when hovering over each author's name, then that revision is skipped.
Any edits that were reverted by the next non-skipped rev will
effectively have never happened, while any edits that were kept by the
next non-skipped rev will be attributed to the author of that rev.