Eliminate the vulcan-forms mergeWithComponents hack
Vulcan-forms has a mechanism where you can pass in some components (in
particular, a submit button and a form-group layout), which it passes
around internaly and uses as part of the form that it builds. This was
implemented via a mechanism that takes an entire copy of the
`Components` table, which we did by creating a proxy-object using
`mergeWithComponents` to create an object that replaces one or two
components while passing the rest through to the global Components
table.

This was a lot of pointless complexity, given that all we actually
wanted to do was replace the Submit button. Narrow the types to disallow
passing in components other than the two that we actually pass, then
drop the whole mechanism in favor of the much simpler
`formComponents?.FormSubmit ??  Components.FormSubmit`.