Add "yarn repl"
This refactors parts of migrate.ts into a shared file scripts/scriptUtil.ts,
to be used in cases where you want to run code in a server-like context (with
a database connection etc) that is not actually a server. It then adds scripts/repl.ts
(and a corresponding target "yarn repl" in package.json).

The idea here is that we will take the various scripts, manual migrations, etc that
currently either go in manualMigrations or are monkeypatched onto `Globals`, and
change the idiom for running them to one which doesn't requite that those things
be imported into regular webservers.