Remove function checkForMemoryLeaks which was using lots of CPU time in prod
We had a check that, every 2s, checks the server-process's memory usage,
and if it's over a threshold, logs a warning and calls a function
`checkForMemoryLeaks`, which checks some global variables for signs of
them having grown larger than they're supposed to. It turns out that
checking those globals was itself a slow operation, and, since it's all
in one uninterruptiple block of CPU time, it noticeably affected P75
latency. And it wasn't actually finding anything to warn about anyways.

Slow the interval from 2s to 10s, and remove the slow part.