Refactor makeEditable away from using global state.
In particular, we now do two things instead of calling makeEditable on a collection:
1. we manually add editable fields to that collection's schema (using the same set of options previously passed to makeEditable) using a new helper function `editableFields`
2. we manually invoke the (refactored) editable callbacks at each necessary stage inside of `createMutator` and `updateMutator`

We also remove all the effectful tracking of what collections have editable fields, what their options are, etc, and replace those with helper functions which get the necessary information on demand.  Currently those rely on `getAllCollections`; hopefully soon that'll be `getAllSchemas` (or something) so that we can avoid pulling collections into the client.