Don't run react-html-email's validator on our stylesheets
This produced a big pile of warnings about potentially unsupported CSS
properties in our styles, for extremely basic CSS properties like
line-height and padding-left. This notion of what CSS is supported by
email clients is very out of date, and we don't want to (and don't in
practice) support email clients that old. Also, the warnings were being
console-logged and never looked at.

Co-Authored-By: Robert Mushkatblat <bobert.mushky@gmail.com>