Fix the Share button in the post triple-dot menu doing nothing
The post triple-dot menu uses `navigator.canShare` to determine whether
it has sharing Intents (ie, whether this is a phone where non-browser
apps have registered intent handlers for sharing URLs). The intended
behavior is that on a phone, clicking the Share button in the post
triple-dot menu sends an intent, while on desktop, it opens a submenu
with Copy Link and other options. However, `navigator.canShare` is a
function, not a boolean, so this tried to use the mobile case on desktop
(which didn't work).