Remove material-UI SwipeableDrawer
On mobile, the left-hand sidebar and right-side notifications used
SwipeableDrawer, which provides a gesture where, on touch/mobile, you
can swipe in from the edge to open them. However in practice, this
gesture doesn't work properly because Android/Chrome and iOS/Safari both
started mapping that gesture to back/forward, which makes it very
frustrating to try to open/close sidebars this way.

Replace usage of SwipeableDrawer with regular Drawer, which preserves
the same animation and modal-background effects, without the swipe-in
gesture.