Make an idiom for type-safe cross-file classname references
One of the obstacles to refactoring is that, when moving components
around, there's a possibility of sneaky references to their stylesheet
that weren't covered by the type system, in the form of styles that
refer to `& .ComponentName-className`. This is particularly an issue
with material-UI components, but also happens elsewhere.

This commit introduces an idiom for doing this in a type-safe way, with
a type-only import. You define styles with `defineStyles` (it must be
the hook version, not the HoC version), export the type of the styles,
then in a stylesheet use

    `.getClassName<StyleType>("ComponentName", "ClassName")`

This typechecks iff the mentioned styles have the right name and have a
class by that name.