Part 1 of removing the Walled Garden GatherTown integration, which has been disused and nonfunctional for awhile
This PR removes all routes, components, client-side usage of the Walled
Garden feature, which was a GatherTown integration we set up in 2020.
This does *not* remove the backend fields on the User object and the
GardenCodes table; that's a second PR, which must not have its
migrations run until after this PR is deployed, because dropping fields
from the User object while they're still in use would cause an outage.