Don't use theme-overrides mechanism for any non-MUI components
The material-UI theming system has a concept of "overrides", where the
theme contains styles identified by component name, that will be merged
with a component's styles. We used this for some forum-gating (of both
EA Forum and Alignment Forum styles) before we had a better idea of what
our forum-gating mechanisms should be, but it's a bad mechanism because
it makes it very easy to look at a component and overlook some styles it
has.

Remove all overrides from our themes, for non-MUI components. (MUI
components are removed from overrides separately, as they get
typescriptified or removed.)