Fix Apollo error when updating cache entry or a query that is in an initial loading state
When you perform an update mutation, we have some code that attempts to
find and update queries that would be affected by the change. However,
if there is a query that would be affected by the change and it's still
in an initial loading state, this crashes with a null-reference
exception.

In practice this doesn't come up often, but one repro is that if you're
logged in as a moderator and you open the notification tray while the
Sunshine Sidebar new users list is loading, it will console-log an error
and fail to clear your notification count.