Prevent apollo-client from double-fetching queries that are in the preloaded cache after SSR
When you load an SSRed page, there's an apollo-client cache included at
the end, containing the results of all queries used in that page, which
is used during hydration. We want queries that are in the cache to not
then make network requests, even if they have a fetch mode of
`cache-and-network` (cache-and-network is a fetch mode that would follow
up with a network request, but which is intended for client-side
navigation, not for SSR).

In order to prevent unnecessary extra network requests, we set the
`disableNetworkFetches` option on apollo-client, and then unset it after
hydration. At some point, this stopped working--most likely with the
React 19 upgrade, but it might have been earlier. This doesn't slow down
the time until the page is displayed, but does waste some bandwidth and
server CPU afterwards. This is probably because "being done with
hydration" isn't a sharp concept anymore with React being more
asynchronous (since React 18).

Fix this by using apollo-client's `ssrForceFetchDelay` option instead,
which does approximately the same thing but leaves the details of when
to reenable fetching to apollo-client, and which seems to work.