Add memoization to CommentsListSection and CommentsTableOfContents
On post pages, after pageload, the PostsPage component rerenders a few
times (for spurious reasons). This causes CommentsListSection and
CommentsTableOfContents to also rerender, which is CPU-intensive if the
page has a large number of comments. Fix that by memoizing.

This wraps loadMore in useMulti in `useStabilizedCallback`, which is
necessary since that function is passed to CommentsListSection. This is
safe because we don't need CommentsListSection (or anything else) to
rerender as a result of it changing.