Fix some non-fatal crashes in SSR that happen during unit tests
`runCrosspostsTests` had this error in its output (without actually
failing the test):
```
[WebServer] ReferenceError: document is not defined
    at Component (/home/runner/work/ForumMagnum/ForumMagnum/packages/lesswrong/components/common/LWDialog.tsx:119:7)
```
This is because LWDialog uses a portal in such a way that it only works
client-sided, and EALoginPopover tries to use it in an SSR context. Fix
by making LWDialog not crash if SSRed (though it will wait til after it's
hydrated to actually display anything).