Remove Material Icons font
The Material Icons font appears to be unused (all of the icons we use
come from SVGs in the repo or in node_modules). Verified by downloading
the external stylesheet, which contained a font name and a CSS class,
and searching the codebase for usages of either, neither of which was
present.