Link-previews can't add side-items to the parent post
This fixes a rare issue where if a post that has an internal link and also
has footnotes inside the text preview, then hovering the internal link
causes extra copies of the footnotes to appear in the sidebar. Which, if
you're hovering over that internal link in a footnote in the sidebar
itself, pushes it down, creating a hover-unhover flicker.

Comes up in https://www.lesswrong.com/posts/Fg8dtE8HHkDoiGcwt/new-feature-support-for-footnotes
(which is a post that deliberately shows off corner-cases of footnotes).