Fix broken radio-buttons
An earlier refactor (vendoring/trimming MUI) broke RadioGroup, which is
used in two places: karma-change notifier settings, and the max karma in
the Subscribe dialog. The issue was that the RadioGroup component would
look at the props of its children to find an associated `value`, and use
this to insert `checked` and `onChanged` props, but the `value` prop
moved to a different place as a result of the radio buttons themselves
being wrapped by `FormControlLabel`.

Fix by tossing the `RadioGroup` component, which does nothing valuable.