Improve the deveoper ergonomics of dark-mode colors
We have a unit test to prevent black-on-black text on dark mode, which
checks styles to make sure that when they contain colors, those colors
differ between light mode and dark mode. This was originally framed as
"colors can only come from the theme's palette". This was later
generalized to allow non-palette colors if they're different between
light and dark mode, ie if you used
  `theme.themeOptions.name === "dark" ? "colorOne" : "colorTwo"`
but this option wasn't mentioned in the unit-test error message, and is
kind of awkward.

Add `theme.dark`, as a shorter substitute for `theme.themeOptions.name`.
Also modify the unit-test's error message to mention the option of
branching on `theme.dark`, and to mention the option of passing the
`{allowNonThemeColors: true}` option.