Reduce rerendering when the current-user object changes
When the current-user object changes, every component that subscribes to
it with useCurrentUser needs to rerender. This is a problem when the
change is something that should be trivial like "update
lastNotificationsCheck" and we're on a post-page with a lot of comments.

Reduce the amount of rerendering that's triggered by opening the
notifications menu and updating lastNotificationsCheck, with a
combination of more memoization, replacing useCurrentUser with
useFilteredCurrentUser, and moving useCurrentUser calls around.

This introduces a new function useFilteredCurrentUser, which uses the
library use-context-selector to make something that subscribes to the
current-user context, but runs a function, and only rerenders when the
result of that function changes.