[WIP] useServerInsertedHtml and server-inserted styles
This commit adds a hook useServerInsertedHtml, with the same effect as
the nextjs function by the same name, and uses it to inject styles into
the page, replacing /allStyles. This makes it so that if a user doesn't
have /allStyles cached, significantly less total data needs to be
downloaded in order to display the page; and, regardless of whether
things are cached or not, browser layout/paint operations are faster
because the page only contains CSS for styles that are actually used,
rather than all styles in the codebase. When components unsuspend, the
corresponding styles (if they are being used for the first time) appear
in the response stream just before the HTML that uses them.
Style-precedence semantics are preserved using a small
outside-the-bundle script that moves styles into place based on their
priority.

Currently known issues:
 * This doesn't play well yet with auto-dark-mode; the injected styles
   are always a concrete theme, rather than using the
   prefers-color-schema media query.
 * Ths client-side JSS insertion/reference-counting doesn't recognize
   the style nodes inserted by SSR, and creates a set of duplicates
   rather than taking ownership of them
 * The theme-switcher hasn't been updated to reflect the changes, if you
   switch theme it will leave the page in a mixed-light-and-dark mode
   until you refresh
 * JSS-to-CSS conversions that run server-side in SSR aren't cached and
   aren't minified (not a huge performance loss, but not ideal)