Background tasks (voided promises) use the `backgroundTask` utility function
In a serverless context, processes may (but don't necessarily) terminate
when the corresponding request finishes, which means that background
tasks need something to indicate that they're running. This PR adds a
function `backgroundTask`, and replaces (most) voided promises with
calls to it. The function doesn't currently do anything other than void
the promise that it receives, but it provides a central place to add the
keep-the-server-running behavior. Also changes the eslint config to
prevent ignoring promises with `void`, to ensure that the
`backgroundTask` wrapper is used.

Note that calls to backgroundTask replace existing voided promises, and
it is _not_ always safe to change their timing, eg to defer them until
the end of the request. In some cases, the background task is receiving
a callback function and will rejoin the main thread. In some cases,
there isn't a request involved at all, eg manual migrations.